Moved package home page from https://github.com/iNamik/lexer.go to https://github.com/iNamik/go_lexer
It turns out that Go's clean functionality can be dangerious if you have a non-derived file with same exact name as the containing package folder.  This is currently a problem with the lexer project with Go's clean removing the /lexer.go/lexer.go src file.

To reduce the potential for this problem in the future, I am moving all of my Go projects from '*.go' to 'go_*'

I also updated references to the iNamik go_container project, which has already been renamed on github.